+++ example/integration_test/badge_test.dart
@@ -5,18 +5,14 @@ import 'package:xue_hua_app_badge/xue_hua_app_badge.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  setUpAll(() async {
-    await XueHuaAppBadge.initialize();
+  test('badge API smoke test', () async {
+    expect(() async => await XueHuaAppBadge.instance.set(0), returnsNormally);
+    expect(() async => await XueHuaAppBadge.instance.set(1), returnsNormally);
+    expect(() async => await XueHuaAppBadge.instance.remove(), returnsNormally);
+    expect(await XueHuaAppBadge.instance.isPermissionGranted(), isA<bool>());
   });
 
-  test('badge API smoke test', () {
-    expect(() => XueHuaAppBadge.set(0), returnsNormally);
-    expect(() => XueHuaAppBadge.set(1), returnsNormally);
-    expect(() => XueHuaAppBadge.remove(), returnsNormally);
-    expect(XueHuaAppBadge.isPermissionGranted(), isA<bool>());
-  });
-
-  test('requestPermission returns bool', () {
-    expect(XueHuaAppBadge.requestPermission(), isA<bool>());
+  test('requestPermission returns bool', () async {
+    expect(await XueHuaAppBadge.instance.requestPermission(), isA<bool>());
   });
 }