+++ example/linux/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,10 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <xue_hua_app_badge/xue_hua_app_badge_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
+  g_autoptr(FlPluginRegistrar) xue_hua_app_badge_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "XueHuaAppBadgePlugin");
+  xue_hua_app_badge_plugin_register_with_registrar(xue_hua_app_badge_registrar);
 }