+++ example/linux/flutter/generated_plugins.cmake
@@ -3,10 +3,10 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  xue_hua_app_badge
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
-  xue_hua_app_badge
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)