+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,9 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <xue_hua_app_badge/xue_hua_app_badge_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
+  XueHuaAppBadgePluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("XueHuaAppBadgePlugin"));
 }