+++ linux/xue_hua_app_badge_plugin_private.h
@@ -1,0 +1,7 @@
+#include <flutter_linux/flutter_linux.h>
+
+#include "include/xue_hua_app_badge/xue_hua_app_badge_plugin.h"
+
+struct _XueHuaAppBadgePlugin {
+  GObject parent_instance;
+};