+++ rust/src/api/mod.rs
@@ -1,3 +1,0 @@
-pub mod frb_handler;
-pub mod badge;
-pub mod simple;